Twelve: Don't be overly protective about permissions
The first thing we do now is requesting the required
permissions and won't load any UI until those are granted
to ensure proper initialization of all the interface
components.

Most of those checks were redunant as they were reachable
only if we already had permissions granted, so simply
do it once.

We don't expose any of those components as-is and all of
them are routed through `MainFragment`.

Permissions can be revoked at runtime (via code or adb)
but this use-case isn't fixable as there isn't any way
to receive updates on such event.